# raft: make Message.Snapshot nullable, halve struct size