--- a/contrib/raftexample/raft_test.go
+++ b/contrib/raftexample/raft_test.go
@@ -37,7 +37,7 @@ func TestProcessMessages(t *testing.T) {
 				{
 					Type: raftpb.MsgSnap,
 					To:   8,
-					Snapshot: raftpb.Snapshot{
+					Snapshot: &raftpb.Snapshot{
 						Metadata: raftpb.SnapshotMetadata{
 							Index: 100,
 							Term:  3,
@@ -53,7 +53,7 @@ func TestProcessMessages(t *testing.T) {
 				{
 					Type: raftpb.MsgSnap,
 					To:   8,
-					Snapshot: raftpb.Snapshot{
+					Snapshot: &raftpb.Snapshot{
 						Metadata: raftpb.SnapshotMetadata{
 							Index: 100,
 							Term:  3,
@@ -74,7 +74,7 @@ func TestProcessMessages(t *testing.T) {
 				{
 					Type: raftpb.MsgSnap,
 					To:   8,
-					Snapshot: raftpb.Snapshot{
+					Snapshot: &raftpb.Snapshot{
 						Metadata: raftpb.SnapshotMetadata{
 							Index: 100,
 							Term:  3,
@@ -95,7 +95,7 @@ func TestProcessMessages(t *testing.T) {
 				{
 					Type: raftpb.MsgSnap,
 					To:   8,
-					Snapshot: raftpb.Snapshot{
+					Snapshot: &raftpb.Snapshot{
 						Metadata: raftpb.SnapshotMetadata{
 							Index: 100,
 							Term:  3,