--- a/server/etcdserver/api/rafthttp/functional_test.go
+++ b/server/etcdserver/api/rafthttp/functional_test.go
@@ -71,7 +71,7 @@ func TestSendMessage(t *testing.T) {
 		{Type: raftpb.MsgAppResp, From: 1, To: 2, Term: 1, Index: 3},
 		{Type: raftpb.MsgVote, From: 1, To: 2, Term: 1, Index: 3, LogTerm: 0},
 		{Type: raftpb.MsgVoteResp, From: 1, To: 2, Term: 1},
-		{Type: raftpb.MsgSnap, From: 1, To: 2, Term: 1, Snapshot: raftpb.Snapshot{Metadata: raftpb.SnapshotMetadata{Index: 1000, Term: 1}, Data: data}},
+		{Type: raftpb.MsgSnap, From: 1, To: 2, Term: 1, Snapshot: &raftpb.Snapshot{Metadata: raftpb.SnapshotMetadata{Index: 1000, Term: 1}, Data: data}},
 		{Type: raftpb.MsgHeartbeat, From: 1, To: 2, Term: 1, Commit: 3},
 		{Type: raftpb.MsgHeartbeatResp, From: 1, To: 2, Term: 1},
 	}