--- a/server/etcdserver/api/rafthttp/snapshot_test.go
+++ b/server/etcdserver/api/rafthttp/snapshot_test.go
@@ -45,7 +45,7 @@ func TestSnapshotSend(t *testing.T) {
 	}{
 		// sent and receive with no errors
 		{
-			m:    raftpb.Message{Type: raftpb.MsgSnap, To: 1},
+			m:    raftpb.Message{Type: raftpb.MsgSnap, To: 1, Snapshot: &raftpb.Snapshot{}},
 			rc:   strReaderCloser{strings.NewReader("hello")},
 			size: 5,
 
@@ -54,7 +54,7 @@ func TestSnapshotSend(t *testing.T) {
 		},
 		// error when reading snapshot for send
 		{
-			m:    raftpb.Message{Type: raftpb.MsgSnap, To: 1},
+			m:    raftpb.Message{Type: raftpb.MsgSnap, To: 1, Snapshot: &raftpb.Snapshot{}},
 			rc:   &errReadCloser{fmt.Errorf("snapshot error")},
 			size: 1,
 
@@ -63,7 +63,7 @@ func TestSnapshotSend(t *testing.T) {
 		},
 		// sends less than the given snapshot length
 		{
-			m:    raftpb.Message{Type: raftpb.MsgSnap, To: 1},
+			m:    raftpb.Message{Type: raftpb.MsgSnap, To: 1, Snapshot: &raftpb.Snapshot{}},
 			rc:   strReaderCloser{strings.NewReader("hello")},
 			size: 10000,
 
@@ -72,7 +72,7 @@ func TestSnapshotSend(t *testing.T) {
 		},
 		// sends less than actual snapshot length
 		{
-			m:    raftpb.Message{Type: raftpb.MsgSnap, To: 1},
+			m:    raftpb.Message{Type: raftpb.MsgSnap, To: 1, Snapshot: &raftpb.Snapshot{}},
 			rc:   strReaderCloser{strings.NewReader("hello")},
 			size: 1,