@ahrtr 2022-10-16
This looks like a valid improvement to me. 

A couple of generic questions:
1. Have you verified that an etcd instance with the new raft (enhanced in this PR) can work with an etcd with the old raft? 
2. Have you done any kinds of benchmark test and how much performance can be improved? 

cc @ptabor @tbg