@nvanbenschoten 2022-10-17
> Have you verified that an etcd instance with the new raft (enhanced in this PR) can work with an etcd with the old raft?

No, not yet. We should perform such verification. Are there cross-version tests in etcd? @serathius mentioned e2e testing, which may include mixed version testing. If no such testing exists, we can piggyback on top of CockroachDB's mixed-version testing to gain more confidence in this change.

> Have you done any kinds of benchmark test and how much performance can be improved?

Also not yet. Are there any such benchmarks in etcd which exercise raft that you'd like to see? If not, I can evaluate this change using the raft benchmarking harness ([rafttoy](https://github.com/nvanbenschoten/rafttoy)) that we developed to measure the performance impact of variations in the use of `etcd/raft`. The harness should also pick up on changes to `etcd/raft` itself.

> If the message is rarely used why change it at all. I know 52% improvement seems great, however if we say rare means ~1/1000 proto messages are snapshot, then this i only 0.05% improvement.

`etcd/raft` exposes a message-passing interface. Every operation (proposal, log append, vote, heartbeat, snapshot, etc.) is encoded into the same `raftpb.Message` struct. [This is](https://github.com/etcd-io/etcd/blob/0ff632c00411c52fdd34d8f9b472915b3ad67ff6/raft/raftpb/raft.proto#L46) the list of all message types. The structs are passed by value into various functions, packed into slices, and shipped over the network.

This commit makes the snapshot header in the message struct nullable, so that non-snapshot messages (the ~999/1000 messages that aren't snapshots) don't need to pay the cost of this large snapshot header either in-memory or over the wire. That reduces the size of the memory representation of a `raftpb.Message` from 264 bytes to 128 bytes (on a 64-bit architecture). It also reduces the proto encoded size of an empty `raftpb.Message` from 30 bytes to 18 bytes.