@ahrtr 2022-10-17
We do have [cluster downgrade test](https://github.com/etcd-io/etcd/blob/main/tests/e2e/cluster_downgrade_test.go), but we don't have mixed version testing, such as 3.5 mixes with 3.6 version. We need to make sure etcd cluster with mixed version can handle raft message correctly. So please add at least one e2e test for this. 

Currently we only have [benchmark](https://github.com/etcd-io/etcd/tree/main/tools/benchmark) tool (FYI. https://github.com/etcd-io/etcd/pull/14394#issuecomment-1229673875). I think we also need to verify the memory & network bandwidth usage probably by the metrics. Does it also make sense to add golang benchmark test ?