@nvanbenschoten 2022-10-21
Thanks for the pointers @ahrtr. I ran some system and micro-benchmarks.

All performance testing was run on a `30vCPU, 128GB, Intel(R) Xeon(R) CPU @ 3.10GHz` GCP instance.

## Results on one-server cluster
```
$ ./bin/etcd --quota-backend-bytes=4300000000

$ go run ./tools/benchmark txn-put --endpoints="http://127.0.0.1:2379" --clients=200 --conns=200 --key-space-size=1000000000 --key-size=128 --val-size=1024  --total=1000000 --rate=40000
```

Results on `main`
```
Summary:
  Total:	31.4938 secs.
  Slowest:	0.0483 secs.
  Fastest:	0.0012 secs.
  Average:	0.0052 secs.
  Stddev:	0.0046 secs.
  Requests/sec:	31752.2840

Response time histogram:
  0.0012 [1]	|
  0.0059 [893343]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0106 [36584]	|∎
  0.0153 [12309]	|
  0.0200 [9395]	|
  0.0247 [40703]	|∎
  0.0295 [4590]	|
  0.0342 [2111]	|
  0.0389 [544]	|
  0.0436 [219]	|
  0.0483 [201]	|

Latency distribution:
  10% in 0.0035 secs.
  25% in 0.0037 secs.
  50% in 0.0039 secs.
  75% in 0.0043 secs.
  90% in 0.0062 secs.
  95% in 0.0195 secs.
  99% in 0.0242 secs.
  99.9% in 0.0341 secs.
```

Results on branch `nvanbenschoten/nilSnapMsg`
```
Summary:
  Total:	31.4111 secs.
  Slowest:	0.0459 secs.
  Fastest:	0.0008 secs.
  Average:	0.0052 secs.
  Stddev:	0.0045 secs.
  Requests/sec:	31835.8783

Response time histogram:
  0.0008 [1]	|
  0.0053 [875383]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0098 [49729]	|∎∎
  0.0143 [14657]	|
  0.0188 [9009]	|
  0.0234 [39721]	|∎
  0.0279 [8964]	|
  0.0324 [1364]	|
  0.0369 [746]	|
  0.0414 [359]	|
  0.0459 [67]	|

Latency distribution:
  10% in 0.0034 secs.
  25% in 0.0036 secs.
  50% in 0.0039 secs.
  75% in 0.0042 secs.
  90% in 0.0064 secs.
  95% in 0.0193 secs.
  99% in 0.0236 secs.
  99.9% in 0.0337 secs.
```

## Results on three-server cluster
```
$ goreman start

$ go run ./tools/benchmark txn-put --endpoints="http://127.0.0.1:2379" --clients=200 --conns=200 --key-space-size=1000000000 --key-size=128 --val-size=1024  --total=1000000 --rate=40000
```

Results on `main`
```
Summary:
  Total:	63.4617 secs.
  Slowest:	0.0578 secs.
  Fastest:	0.0034 secs.
  Average:	0.0123 secs.
  Stddev:	0.0049 secs.
  Requests/sec:	15757.5301

Response time histogram:
  0.0034 [1]	|
  0.0088 [94433]	|∎∎∎∎∎
  0.0142 [731650]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0197 [73329]	|∎∎∎∎
  0.0251 [66848]	|∎∎∎
  0.0306 [23264]	|∎
  0.0360 [6692]	|
  0.0415 [2576]	|
  0.0469 [892]	|
  0.0524 [121]	|
  0.0578 [194]	|

Latency distribution:
  10% in 0.0089 secs.
  25% in 0.0097 secs.
  50% in 0.0107 secs.
  75% in 0.0125 secs.
  90% in 0.0197 secs.
  95% in 0.0234 secs.
  99% in 0.0308 secs.
  99.9% in 0.0423 secs.
```

Results on branch `nvanbenschoten/nilSnapMsg`
```
Summary:
  Total:	63.4009 secs.
  Slowest:	0.1893 secs.
  Fastest:	0.0033 secs.
  Average:	0.0123 secs.
  Stddev:	0.0055 secs.
  Requests/sec:	15772.6380

Response time histogram:
  0.0033 [1]	|
  0.0219 [926219]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0405 [72671]	|∎∎∎
  0.0591 [909]	|
  0.0777 [0]	|
  0.0963 [0]	|
  0.1149 [0]	|
  0.1335 [0]	|
  0.1521 [0]	|
  0.1707 [0]	|
  0.1893 [200]	|

Latency distribution:
  10% in 0.0088 secs.
  25% in 0.0096 secs.
  50% in 0.0107 secs.
  75% in 0.0123 secs.
  90% in 0.0199 secs.
  95% in 0.0238 secs.
  99% in 0.0307 secs.
  99.9% in 0.0411 secs.
```

These tests show stable results and very little difference across branches. Maybe a 0.1% increase in throughput, but I did not run enough iterations to determine whether that difference is in the noise. That's mostly expected, given the size of the benchmark, its interaction with disk and network IO, and how targeted this change is.

## Results on microbenchmarks

Microbenchmarks show a statistically significant and sizable improvement in proposal latency and memory allocations (count and size).

```
name               old time/op    new time/op    delta
Proposal3Nodes-30    3.86µs ±16%    3.69µs ±23%   -4.37%  (p=0.000 n=100+100)

name               old alloc/op   new alloc/op   delta
Proposal3Nodes-30    4.75kB ±39%    3.82kB ±73%  -19.58%  (p=0.000 n=100+100)

name               old allocs/op  new allocs/op  delta
Proposal3Nodes-30      20.4 ±18%      18.0 ±33%  -11.66%  (p=0.000 n=100+100)
```