@ahrtr 2022-10-23
Thanks @nvanbenschoten for the feedback, which looks good. Usually messages do not include snapshot, so it doesn't make sense to always get the snapshot header included in each message. This PR overall looks goo to me.  cc @ptabor and @tbg to double check.

We also need to add an mixed version testing (the main/3.6 mix with the previous version/3.5). Please let me know if you need help on this.