@serathius 2022-10-23
I had idea how to incorporate mixed version testing into common tests by just adding new test scenarios. 
PTAL https://github.com/etcd-io/etcd/pull/14610